+++ lightsim2grid/tests/test_Deactivatedbus.py
@@ -13,12 +13,16 @@ from lightsim2grid.gridmodel import init_from_pandapower
 import pandapower.networks as pn
 import pandapower as pp
 import warnings
-from test_GridModel import BaseTests
+from lightsim2grid.tests.test_GridModel_pandapower import BaseTests
 import pdb
+from global_var_tests import MAX_PP_DATAREADER_NOT_BROKEN, CURRENT_PP_VERSION
 
 
 class MakeACTestsDisco(BaseTests, unittest.TestCase):
     def setUp(self):
+        if CURRENT_PP_VERSION > MAX_PP_DATAREADER_NOT_BROKEN:
+            self.skipTest("Test not correct: pp changed the way it computed trafo params")
+        
         self.net = pn.case118()
         self.last_real_bus = self.net.bus.shape[0]
         pp.create_bus(self.net, vn_kv=self.net.bus["vn_kv"][0])