+++ CHANGELOG.rst
@@ -34,7 +34,8 @@ TODO: integration test with pandapower (see `pandapower/contingency/contingency.
 - [DEPRECATED] python 3.7 builds will no longer be available
 - [FIXED] a bug in the import of the grid from pypowsybl
 - [FIXED] a bug with phase shifters in case the tap was on the low voltage side
-- [ADDED] compatibility with python 3.14
+- [FIXED] a bug with active shunt values (wrong sign in the cpp part)
+- [ADDED] compatibility with python 3.14 and python 3.14 build
 - [IMPROVED] the way to initialize the transformers from pypowsybl
 - [IMPROVED] possibility to load grid with phase shifters from pypowsybl